+++ scripts/goto_session.sh
@@ -5,7 +5,7 @@ CURRENT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 main() {
 	# displayes tmux session list
 	tmux run-shell "$CURRENT_DIR/list_sessions.sh"
-	# invokes goto command prompt
-	"$CURRENT_DIR/show_goto_prompt.sh"
+	# goto command prompt
+	tmux command -p session: "run '$CURRENT_DIR/switch_or_loop.sh %1'"
 }
 main