+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - handle "promote pane" feature bug
 - add `.gitattributes` file so that end of line chracter is always set to `\n`
   on cygwin
+- speed up 'goto session' prompt
 
 ### v2.3.0, 2015-06-19
 - split 'goto session' script into 2 files, make it simpler